requirements.txt
tensorflow-gpu==1.15
numpy>=1.9
tqdm>=4.2
opencv-python>=3.1
moviepy==1.0.3
protobuf==3.20.1
python-dotenv
minio==7.1.12
jupyter==1.0.0